# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: XJU Lab API
  version: 1.0.0
  description: Session-authenticated API for the XJU Lab collaboration platform.
servers:
  - url: /api/v1
...
paths:
...
  /members:
    get:
      tags: [members]
      parameters:
        - { name: page, in: query, schema: { type: integer, minimum: 1, default: 1 } }
        - { name: pageSize, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
      responses: { '200': { description: Authorized member directory, content: { application/json: { schema: { $ref: '#/components/schemas/MemberPage' } } } } }
...
  /projects/{projectId}:
    patch:
      tags: [projects]
      parameters: [{ name: projectId, in: path, required: true, schema: { type: string, format: uuid } }, { $ref: '#/components/parameters/IfMatchVersion' }]
      requestBody: { required: true, content: { application/json: { schema: { type: object, required: [title, status], properties: { title: { type: string, maxLength: 160 }, description: { type: string, maxLength: 5000 }, status: { type: string, enum: [ACTIVE, ARCHIVED] } } } } } }
      responses: { '200': { description: Versioned project update or archive }, '409': { $ref: '#/components/responses/ApiError' } }
  /projects/{projectId}/members:
    get:
      tags: [projects]
      parameters: [{ name: projectId, in: path, required: true, schema: { type: string, format: uuid } }]
      responses: { '200': { description: Members of a project the caller can access } }
  /projects/{projectId}/members/{memberId}:
    put:
      tags: [projects]
      parameters: [{ name: projectId, in: path, required: true, schema: { type: string, format: uuid } }, { name: memberId, in: path, required: true, schema: { type: string, format: uuid } }]
      requestBody: { required: true, content: { application/json: { schema: { type: object, required: [role], properties: { role: { type: string, enum: [MEMBER] } } } } } }
      responses: { '200': { description: Project member added }, '404': { $ref: '#/components/responses/ApiError' } }
    delete:
      tags: [projects]
      parameters: [{ name: projectId, in: path, required: true, schema: { type: string, format: uuid } }, { name: memberId, in: path, required: true, schema: { type: string, format: uuid } }]
      responses: { '204': { description: Project member removed }, '409': { $ref: '#/components/responses/ApiError' } }
  /projects/{projectId}/milestones:
    get:
      tags: [projects]
      parameters: [{ name: projectId, in: path, required: true, schema: { type: string, format: uuid } }]
      responses: { '200': { description: Project milestones visible to the caller } }
    post:
      tags: [projects]
      parameters: [{ name: projectId, in: path, required: true, schema: { type: string, format: uuid } }]
      requestBody: { required: true, content: { application/json: { schema: { type: object, required: [title], properties: { title: { type: string, maxLength: 200 }, dueDate: { type: string, format: date, nullable: true } } } } } }
      responses: { '200': { description: Milestone created }, '404': { $ref: '#/components/responses/ApiError' } }
  /projects/{projectId}/milestones/{milestoneId}:
    patch:
      tags: [projects]
      parameters: [{ name: projectId, in: path, required: true, schema: { type: string, format: uuid } }, { name: milestoneId, in: path, required: true, schema: { type: string, format: uuid } }, { $ref: '#/components/parameters/IfMatchVersion' }]
      requestBody: { required: true, content: { application/json: { schema: { type: object, required: [title, completed], properties: { title: { type: string, maxLength: 200 }, dueDate: { type: string, format: date, nullable: true }, completed: { type: boolean } } } } } }
      responses: { '200': { description: Milestone updated }, '409': { $ref: '#/components/responses/ApiError' } }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: SESSION
    printerAgentToken:
      type: http
      scheme: bearer
  parameters:
    IfMatchVersion:
      name: If-Match-Version
      in: header
      required: true
      schema: { type: integer, minimum: 1 }
  schemas:
...
    MemberPage:
      allOf:
        - $ref: '#/components/schemas/PageEnvelopeBase'
        - type: object
          properties:
            items: { type: array, items: { type: object, properties: { id: { type: string, format: uuid }, displayName: { type: string }, direction: { type: string, nullable: true }, cohort: { type: integer, nullable: true }, studentNumber: { type: string, nullable: true }, contact: { type: string, nullable: true }, version: { type: integer } } } }
    PageEnvelopeBase:
      type: object
      required: [items, total, page, pageSize]
      properties:
        items: { type: array, items: { type: object } }
        total: { type: integer, format: int64 }
        page: { type: integer }
        pageSize: { type: integer }
...
  responses:
    ApiError:
      description: Consistent request error
      content:
        application/json:
          schema:
            type: object
            required: [code, message, requestId]
            properties:
              code: { type: string }
              message: { type: string }
              requestId: { type: string }
              fieldErrors: { type: object, additionalProperties: { type: string } }
security:
  - sessionCookie: []